[8.0] Add a config flag to experiment with some work item prioritization changes in cases involving a lot of sync-over-async (#103984)
* [8.0] Add a config flag to experiment with some work item prioritization changes in cases involving a lot of sync-over-async

Some services that use a lot of sync-over-async were seen to experience stalls due to some priority inversion issues in work items that get queued to the thread pool. For instance, a work item W1 queues another work item W2 to the global queue and blocks waiting for a task to complete, where W2 would need to run in order to complete the task, but W2 is queued behind a number of other work items that operate like W1, and this sometimes leads to long-duration stalls. This change adds an experimental config option that when enabled, enqueues some kinds of work items to a new low-priority global queue that is checked after all other global queues. This was seen to help in some cases.