JIT: Account for mixed-enregistered locals when zeroing without block-init (#104593)
Locals that are in registers at the beginning of a function do not need
to have their stack home zeroed. `genFnProlog` already skips these
locals when computing the range of bytes to zero; however,
`genZeroInitFrame` was not doing the same in the non-block init case,
which does not make use of the range computed by `genFnProlog`. This
could cause an unbounded amount of (unnecessary) codegen during
zero-initing, which is not legal to have in the prolog.

Fix #104570