SPMI: Add a buffer to FileWriter
We are spending a significant amount of time writing the details CSV
file now that the list of metadata has grown quite a bit. This was
especially noticeable during asmdiffs where we write a column for both
the base and diff JIT for each metadata key.

A lot of that cost was coming down to the fact that `FileWriter` was
unbuffered. This adds a buffer to it to cut down on the number of
syscalls performed.

Ideally we would probably switch to std::ostream now that using STL
should be doable, but I'm not adventurous enough to make that change
right now.