fix warnings in coreclr/jit/compiler.cpp for non-Windows targets
This change was extracted from https://github.com/dotnet/runtimelab/pull/2614,
which includes various fixes to enable building on non-Windows systems.  We're
in the process of upstreaming the parts of that PR which are not specific to
NativeAOT-LLVM, and this is the latest such change.

Note that I left a TODO comment in `DisplayNowayAssertMap` since I'm not sure
how to print a `LPCWSTR` using `fprintf` on non-Windows systems, but it's clear
that the existing code is not portable given the difference between `wchar_t` on
Windows (where it's 16 bits) and other systems (where it's usually 32 bits).

Signed-off-by: Joel Dice <joel.dice@fermyon.com>