Fix SSP restoring in edge cases
There are edge cases when the SSP restoring for continuation after a
catch handler completes doesn't work correctly. The problem is caused by
the fact that we scan for the Rip of the frame handling the exception on
the shadow stack to find where to restore it, and in those edge cases,
the same address can be there multiple times and the first occurence is
not the right one. For example, when an exception is thrown from a catch
handler, it escapes the handler and the handler for the escaped
exception is in the same method as the one that invoked the handler.

This change fixes it by finding the SSP of the first managed frame where
we search for the handler and then updating the SSP with every unwind.
The SSP is stored in the REGDISPLAY. So when we reach the
CallCatchFunclet, the REGDISPLAY contains the SSP to restore.

There was also one more issue with restoring the SSP. I turned out that
the incsspq instruction uses only the lowest 8 bits of the argument to
increment the SSP, so the ClrRestoreNonVolatileContextWorker needs to
have a loop that repeats that instruction in case we need to move it by
more than 255 slots.