JIT: Remove `GTF_IS_IN_CSE`
This flag is just a convoluted way to pass an argument through to a
bunch of methods from inside CSE. That's because CSE tries to reuse
`gtExtractSideEffList` even though it needs something more capable that
considers CSE defs and CSE uses as well.

Remove the flag in favor of an `ignoreCctors` flag in the side effect
checking functions; then, additionally add a CSE-specific version of
`gtExtractSideEffList` called `optExtractSideEffectsForCSE` which
handles side effects and also CSE defs/uses. This does result in a
slight amount of duplication, but I think that's beneficial over the
convoluted logic before.