[mono] Fix aot compiler crash when checking for method signature
This fixes a problem with the aot compiler where it tries to get a generic parameter count from a method, but instead crashes because the reference is null.

The problem stemmed from loading an assembly that had references to other assemblies that were not found. The aot compiler eventually figures out it can't do anything and exits. The crash occurred in between the validation process.

Fixes https://github.com/dotnet/runtime/issues/103901