ILLink: Fix order dependent logic for DAM on types (#104701)
When applying DynamicallyAccessedMembers annotations to a type, we
need to ensure that the base annotations are also applied to the base
type.

Previously this logic was determining which annotations to apply to
the base type (with the warning originating from the derived type)
based on whether the base annotations were already applied. This is
incorrect as it introduces an order dependence. Instead the base
annotations should use the base type as the origin.