PR feedback
combined_limit is now synchronized in GcEnumAllocContexts instead of RestartEE.
This requires the GC being constrained in how it updates the alloc_ptr and alloc_limit. No GC behavior changed,
in practice, but the constraints are now part of the EE<->GC contract so that we can rely on them in the EE code.