JIT: Unify LSRA call arg building (#104051)
This unifies the building of uses of call args across all our targets,
and removes accesses of the old ABI information.

LSRA had a bunch of assert code to compare the `PUTARG_REG` nodes to the
ABI info. I don't think this is particularly valuable given that we
create `PUTARG_REG` nodes in one centralized place in lowering now that
follows the same pattern as the new unified code here does.