JIT: Check return value of `InitializeCursors` (#105017)
We were missing checking the return value of `InitializeCursors`.
Typically this would just cause us to end up with an empty set of
cursors as we abort the search as soon as we fail to create one cursor,
and the cursor we would typically fail to create is always the first
one. Running the next logic with an empty set of cursors causes no ill
effects, so this has no diffs.