JIT: Remove `GTF_IND_INVARIANT` and `GTF_IND_NONFAULTING` flags checking (#104976)
These flags are strictly optimizations. Having them required to be set for
certain indirs based on context of the operand introduces IR invariants that are
annoying to work with since it suddenly becomes necessary for all
transformations to consider "did we just introduce this IR shape?". Instead,
handle these patterns during morph as the optimization it is and remove the
strict flags checking from `fgDebugCheckFlags`.

Also fix `HandleKindDataIsInvariant` which returned true in some questionable
cases, and remove some unnecessary indir flags that morph was setting for
`GTF_ICON_OBJ_HDL`.