Fix feature switch IL when feature attributes are removed (#104995)
When removing `FeatureSwitchDefinition` and `FeatureGuard`
attributes, we need to ensure that the body rewriting logic still
respects these attributes even though the rewriting happens after
attribute removal. Caching the substituted constant value fixes
this.

More detail:

When feature switches or feature guards are substituted with a
constant, the IL rewriting of the feature property happens after
SweepStep, during CodeRewriterStep. At this point, any attributes
marked for removal by RemoveAttributeInstancesAttribute have
already been swept. When removing
`FeatureSwitchDefinitionAttribute` and `FeatureGuardAttribute`,
as we do with AggressiveAttributeTrimming, this means that
CodeRewriterStep fails to substitute a constant for attributed
feature check or feature guard properties.

Normally, the attributed feature properties would be removed
entirely when substituted, so CodeRewriterStep wouldn't even
visit the property. However, it's possible that the caller of the
property was a `copy` assembly where the callsite can't be
rewritten to return a constant.

When this happens, we can end up with a property that's not
stubbed out during CodeRewriterStep, even though we didn't mark a
callee of the property (because MarkStep did treat the property
as a constant and didn't mark its instructions). When trying to
write out the property's IL this results in a failure because we
see a reference to a method that was removed.

Fixes https://github.com/dotnet/runtime/issues/104945