[mono][aot] Emit marhsalling exception when dealing with MONO_MARSHAL_CONV_OBJECT_IUNKNOWN
This change prevents the aot compiler from erroring out when dealing with COM types that were not marked with MarshalAs attributes. In most scenarios that we support, we want to allow pinvokes to aot compile as in cases like anything COM interop, will end up erroring out if you try to use it.

Fixes https://github.com/dotnet/runtime/issues/104463