Added assertion
An interface name comes from a scope in a valid IPv6 address. There'll always be at least three characters (::%) in front of the scope, so taking the length of an interface name which stripping these three leading characters away and adds a single null terminator will never overflow an integer.