JIT: Check for potential store-to-load forwarding before reordering ldr -> ldp
Very targeted fix for #93401 and #101437: before reordering two
indirections, check if there is a potential store in the same loop that
looks like it could end up being a candidate for store-to-load
forwarding into one of those indirections. Some hardware does not handle
store-to-load forwarding with the same fidelity when `stp`/`ldp` is
involved compared to multiple `str`/`ldr`.

If we detect the situation then avoid doing the reordering.