Make Config generator test for downcasts
We were hitting compiler errors when the generator emitted test casts
for value types.  Since those can never be true (value types cannot be
derived from, and the compiler can see if the cast will succeed or not).

Fix this by only doing a test cast when we are trying to do a runtime
downcast.