JIT: Check for potential store-to-load forwarding before reordering ldr -> ldp (#105695)
Very targeted fix for #93401 and #101437: before reordering two indirections,
check if there is a potential store in the same loop that looks like it could
end up being a candidate for store-to-load forwarding into one of those
indirections. Some hardware does not handle store-to-load forwarding with the
same fidelity when `stp`/`ldp` is involved compared to multiple `str`/`ldr`.

The detection is done by a graph walk that starts at the indirection and then
walks backwards until it finds a store that would reach the indirection. The
walk is limited to stay within the same loop as the indirections, and also
limited by a budget of 100 nodes visited (for a large loop we expect this to not
be as important).

If we detect the situation then avoid doing the reordering.

Fix #93401 
Fix #101437