Initialize RuntimeEventSource as part of coreclr_initialize instead of only when running an app (#105902)
We were only initializing `RuntimeEventSource` before running an application. This meant that native hosting scenarios - COM, C++/CLI, custom - would never initialize it. As a result, diagnostic tracing tools - like `dotnet-counters` - did not work properly.

This moves the call to `RuntimeEventSource.Initialize` out of `StartupHookProvider.ManagedStartup` and to be part of `coreclr_initialize`.